+++ omnivox-piper-sys/wrapper.h
@@ -1,1 +1,0 @@
-#include "piper_bridge.h"